refactor(scraper): modularize scraper into package
