+++ como_voto_scraper/__init__.py
@@ -1,0 +1,19 @@
+from .core import classify_bloc
+from .db import ConsolidatedDB, VOTE_DECODE, VOTE_ENCODE
+from .hcdn import scrape_diputados, scrape_hcdn_votacion
+from .photos import scrape_diputados_photos, scrape_senadores_photos
+from .runner import main
+from .senado import scrape_senadores
+
+__all__ = [
+    "ConsolidatedDB",
+    "VOTE_DECODE",
+    "VOTE_ENCODE",
+    "classify_bloc",
+    "main",
+    "scrape_diputados",
+    "scrape_hcdn_votacion",
+    "scrape_senadores",
+    "scrape_diputados_photos",
+    "scrape_senadores_photos",
+]
